Refactor: 뱃지에 뱃지 이름 칼럼 추가 / 뱃지 id로 뱃지 정보 조회하는 기능 추가
