+++ Pick/src/main/java/com/nob/pick/badge/query/dto/BadgeQueryDTO.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 public class BadgeQueryDTO {
 	private int id;
 	private String name;
-	private Integer requirement;
+	private int requirement;
 	private int advantage;
 	private String description;
 	private Boolean isDeleted;
