+++ Pick/src/main/java/com/nob/pick/common/client/MemberClient.java
@@ -1,4 +1,4 @@
-package com.nob.pick.dailymission.command.infrastructure;
+package com.nob.pick.common.client;
 
 import java.util.List;
 
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 import com.nob.pick.common.dto.MemberResponse;
 
-@FeignClient(name = "pick-member-service")
+@FeignClient(name = "pick-member-service-challenge")  // 추후 'pick-member-service'로 변경
 public interface MemberClient {
 	@GetMapping("/api/members/id/{id}")  // 개별 회원 ID 가져오기
 	MemberResponse getMember(@PathVariable("id") Long id);
