+++ Pick/src/main/java/com/nob/pick/common/dto/MemberResponse.java
@@ -1,0 +1,10 @@
+package com.nob.pick.common.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class MemberResponse {
+	private Long id;
+}
